requirements.txt
streamlit
google-generativeai
google-auth
beautifulsoup4
requests